fix(sessions): keep reset-boundary planning off the sqlite write cycle
Moving append into session-reset-boundary-event.ts pulled sqlite-read into
the planner leaf and created a madge cycle through the read fence. Header
emptiness also hydrated stored payloads, so the race tests saw a body.

Keep the planner as a leaf. Append lives on the sqlite write path and
probes seq existence instead of loading event_json.
