+++ src/config/sessions/session-reset-boundary-event.ts
@@ -1,9 +1,4 @@
 import { randomUUID } from "node:crypto";
-import type { OpenClawAgentDatabase } from "../../state/openclaw-agent-db.js";
-import { loadTranscriptEventsFromDatabase } from "./session-accessor.sqlite-read.js";
-import type { ResolvedTranscriptScope } from "./session-accessor.sqlite-scope.js";
-import { appendTranscriptEventsInTransaction } from "./session-accessor.sqlite-transcript-store.js";
-import { createSessionTranscriptHeader } from "./transcript-header.js";
 import { selectRecentUserAssistantReplayRecords } from "./transcript-replay.js";
 import { selectSessionTranscriptLeafControlledPath } from "./transcript-tree.js";
 
@@ -72,41 +67,6 @@ function projectLatestBoundaryWindow(entries: readonly unknown[]): unknown[] {
           return role === "user" || role === "assistant";
         });
   return [...kept, ...entries.slice(boundaryIndex + 1)];
-}
-
-/**
- * Appends a reset boundary, prepending a canonical session header only when
- * the transcript is empty. Fresh /new resets otherwise persist the reset as
- * seq 0 and later turns fail the runtime legacy-transcript assertion.
- * Nonempty headerless transcripts stay headerless so doctor can rewrite the
- * header at seq 0.
- */
-export function appendSessionResetBoundaryEventsInTransaction(
-  database: OpenClawAgentDatabase,
-  scope: ResolvedTranscriptScope,
-  request: SessionResetBoundaryRequest,
-  options: { cwd?: string } = {},
-): number {
-  const storedEvents = loadTranscriptEventsFromDatabase(database, scope.sessionId);
-  const events = loadTranscriptEventsFromDatabase(database, scope.sessionId, {
-    projection: "reset-boundary",
-  });
-  const event = buildSessionResetBoundaryEvent({
-    events,
-    ...request,
-  });
-  // Only physically empty transcripts need a header here. Reset-boundary
-  // projection can hide existing rows, and a nonempty headerless transcript
-  // must stay headerless so doctor can rewrite the header at seq 0.
-  const batch =
-    storedEvents.length === 0
-      ? [createSessionTranscriptHeader({ cwd: options.cwd, sessionId: scope.sessionId }), event]
-      : [event];
-  const appended = appendTranscriptEventsInTransaction(database, scope, batch);
-  if (appended !== batch.length) {
-    throw new Error(`Failed to append reset boundary for ${scope.sessionKey}`);
-  }
-  return appended;
 }
 
 export function buildSessionResetBoundaryEvent(
